refactor(module:tooltip,popover,popconfirm): refactor (#4721)
* chore(module:tooltip,popover,popconfirm): refactor

chore: refactor

chore: refactor

chore: merge redundant files

chore: remove redundant files

fix: path

fix: remove unused method

fix: remove useless ng-content

fix: constructor

* feat(module:tooltip): close when title is null
