+++ components/popconfirm/public-api.ts
@@ -6,6 +6,5 @@
  * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
  */
 
-export * from './nz-popconfirm.component';
-export * from './nz-popconfirm.directive';
-export * from './nz-popconfirm.module';
+export * from './popconfirm';
+export * from './popconfirm.module';
