--- components/popover/nz-popover.component.html
@@ -1,36 +1,0 @@
-<ng-content></ng-content>
-<ng-template
-  #overlay="cdkConnectedOverlay"
-  cdkConnectedOverlay
-  nzConnectedOverlay
-  [cdkConnectedOverlayOrigin]="origin"
-  [cdkConnectedOverlayHasBackdrop]="_hasBackdrop"
-  (backdropClick)="hide()"
-  (detach)="hide()"
-  (positionChange)="onPositionChange($event)"
-  [cdkConnectedOverlayPositions]="_positions"
-  [cdkConnectedOverlayOpen]="_visible"
->
-  <div
-    class="ant-popover"
-    [ngClass]="_classMap"
-    [ngStyle]="nzOverlayStyle"
-    [@.disabled]="noAnimation?.nzNoAnimation"
-    [nzNoAnimation]="noAnimation?.nzNoAnimation"
-    [@zoomBigMotion]="'active'"
-  >
-    <div class="ant-popover-content">
-      <div class="ant-popover-arrow"></div>
-      <div class="ant-popover-inner" role="tooltip">
-        <div>
-          <div class="ant-popover-title" *ngIf="title">
-            <ng-container *nzStringTemplateOutlet="title">{{ title }}</ng-container>
-          </div>
-          <div class="ant-popover-inner-content">
-            <ng-container *nzStringTemplateOutlet="content">{{ content }}</ng-container>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</ng-template>
