--- components/popover/nz-popover.component.ts
@@ -1,40 +1,0 @@
-/**
- * @license
- * Copyright Alibaba.com All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
- */
-
-import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Host, Optional, ViewEncapsulation } from '@angular/core';
-
-import { NzNoAnimationDirective, zoomBigMotion } from 'ng-zorro-antd/core';
-import { isTooltipEmpty, NzToolTipComponent } from 'ng-zorro-antd/tooltip';
-
-@Component({
-  selector: 'nz-popover',
-  exportAs: 'nzPopoverComponent',
-  animations: [zoomBigMotion],
-  templateUrl: './nz-popover.component.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-  encapsulation: ViewEncapsulation.None,
-  preserveWhitespaces: false,
-  styles: [
-    `
-      .ant-popover {
-        position: relative;
-      }
-    `
-  ]
-})
-export class NzPopoverComponent extends NzToolTipComponent {
-  _prefix = 'ant-popover-placement';
-
-  constructor(cdr: ChangeDetectorRef, @Host() @Optional() public noAnimation?: NzNoAnimationDirective) {
-    super(cdr, noAnimation);
-  }
-
-  protected isEmpty(): boolean {
-    return isTooltipEmpty(this.title) && isTooltipEmpty(this.content);
-  }
-}
