+++ components/popover/popover.ts
@@ -1,0 +1,117 @@
+/**
+ * @license
+ * Copyright Alibaba.com All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
+ */
+
+import {
+  ChangeDetectionStrategy,
+  ChangeDetectorRef,
+  Component,
+  ComponentFactory,
+  ComponentFactoryResolver,
+  Directive,
+  ElementRef,
+  Host,
+  Input,
+  Optional,
+  Renderer2,
+  ViewContainerRef,
+  ViewEncapsulation
+} from '@angular/core';
+
+import { NzNoAnimationDirective, NzTSType, zoomBigMotion } from 'ng-zorro-antd/core';
+import { isTooltipEmpty, NzTooltipBaseDirective, NzToolTipComponent, NzTooltipTrigger } from 'ng-zorro-antd/tooltip';
+
+@Directive({
+  selector: '[nz-popover]',
+  exportAs: 'nzPopover',
+  host: {
+    '[class.ant-popover-open]': 'isTooltipComponentVisible'
+  }
+})
+export class NzPopoverDirective extends NzTooltipBaseDirective {
+  @Input('nzPopoverTitle') specificTitle: NzTSType;
+  @Input('nzPopoverContent') specificContent: NzTSType;
+  @Input('nz-popover') directiveNameTitle: NzTSType | null;
+  @Input('nzPopoverTrigger') specificTrigger: NzTooltipTrigger;
+  @Input('nzPopoverPlacement') specificPlacement: string;
+
+  componentFactory: ComponentFactory<NzPopoverComponent> = this.resolver.resolveComponentFactory(NzPopoverComponent);
+
+  constructor(
+    elementRef: ElementRef,
+    hostView: ViewContainerRef,
+    resolver: ComponentFactoryResolver,
+    renderer: Renderer2,
+    @Host() @Optional() public noAnimation?: NzNoAnimationDirective
+  ) {
+    super(elementRef, hostView, resolver, renderer, noAnimation);
+  }
+}
+
+@Component({
+  selector: 'nz-popover',
+  exportAs: 'nzPopoverComponent',
+  animations: [zoomBigMotion],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  encapsulation: ViewEncapsulation.None,
+  preserveWhitespaces: false,
+  template: `
+    <ng-template
+      #overlay="cdkConnectedOverlay"
+      cdkConnectedOverlay
+      nzConnectedOverlay
+      [cdkConnectedOverlayOrigin]="origin"
+      [cdkConnectedOverlayHasBackdrop]="_hasBackdrop"
+      (backdropClick)="hide()"
+      (detach)="hide()"
+      (positionChange)="onPositionChange($event)"
+      [cdkConnectedOverlayPositions]="_positions"
+      [cdkConnectedOverlayOpen]="_visible"
+    >
+      <div
+        class="ant-popover"
+        [ngClass]="_classMap"
+        [ngStyle]="nzOverlayStyle"
+        [@.disabled]="noAnimation?.nzNoAnimation"
+        [nzNoAnimation]="noAnimation?.nzNoAnimation"
+        [@zoomBigMotion]="'active'"
+      >
+        <div class="ant-popover-content">
+          <div class="ant-popover-arrow"></div>
+          <div class="ant-popover-inner" role="tooltip">
+            <div>
+              <div class="ant-popover-title" *ngIf="title">
+                <ng-container *nzStringTemplateOutlet="title">{{ title }}</ng-container>
+              </div>
+              <div class="ant-popover-inner-content">
+                <ng-container *nzStringTemplateOutlet="content">{{ content }}</ng-container>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </ng-template>
+  `,
+  styles: [
+    `
+      .ant-popover {
+        position: relative;
+      }
+    `
+  ]
+})
+export class NzPopoverComponent extends NzToolTipComponent {
+  _prefix = 'ant-popover-placement';
+
+  constructor(cdr: ChangeDetectorRef, @Host() @Optional() public noAnimation?: NzNoAnimationDirective) {
+    super(cdr, noAnimation);
+  }
+
+  protected isEmpty(): boolean {
+    return isTooltipEmpty(this.title) && isTooltipEmpty(this.content);
+  }
+}
