--- components/tooltip/nz-tooltip.component.html
@@ -1,29 +1,0 @@
-<ng-content></ng-content>
-<ng-template
-  #overlay="cdkConnectedOverlay"
-  cdkConnectedOverlay
-  nzConnectedOverlay
-  [cdkConnectedOverlayOrigin]="origin"
-  [cdkConnectedOverlayOpen]="_visible"
-  [cdkConnectedOverlayHasBackdrop]="_hasBackdrop"
-  [cdkConnectedOverlayPositions]="_positions"
-  (backdropClick)="hide()"
-  (detach)="hide()"
-  (positionChange)="onPositionChange($event)"
->
-  <div
-    class="ant-tooltip"
-    [ngClass]="_classMap"
-    [ngStyle]="nzOverlayStyle"
-    [@.disabled]="noAnimation?.nzNoAnimation"
-    [nzNoAnimation]="noAnimation?.nzNoAnimation"
-    [@zoomBigMotion]="'active'"
-  >
-    <div class="ant-tooltip-content">
-      <div class="ant-tooltip-arrow"></div>
-      <div class="ant-tooltip-inner">
-        <ng-container *nzStringTemplateOutlet="title">{{ title }}</ng-container>
-      </div>
-    </div>
-  </div>
-</ng-template>
